+++ app/helpers/evaluation_helper.rb
@@ -1,8 +1,28 @@
 # coding: utf-8
 
 module EvaluationHelper
-  def row_class(numbers, ground_truth)
-    numbers.include?(ground_truth) ? 'success' : 'danger'
+  def progress(evaluation)
+    case evaluation.state
+    when 'completed'
+      '完了'
+    when 'error'
+      'エラー'
+    else
+      if evaluation.data.empty?
+        '0%完了'
+      else
+        completed_data_size = evaluation.data.to_a.count do |datum|
+          datum.prediction_results.present?
+        end
+        "#{(100 * completed_data_size / evaluation.data.size.to_f).round(0)}%完了"
+      end
+    end
+  end
+
+  def row_class(numbers, datum)
+    return 'warning' if datum.prediction_results.empty?
+
+    numbers.include?(datum.ground_truth) ? 'success' : 'danger'
   end
 
   def span_color(number, ground_truth)